refactor: move automatic consolidation services
