+++ packages/app/package.json
@@ -25,6 +25,7 @@
     "dependencies": {
         "@budgie/ai": "workspace:^",
         "@budgie/bank-sync": "workspace:*",
+        "@budgie/consolidation": "workspace:*",
         "@budgie/contracts": "workspace:^",
         "@budgie/logger": "workspace:*",
         "@expo/fingerprint": "~0.16.7",
