+++ packages/consolidation/src/auto/interface/consolidation-candidate-dependencies.interface.ts
@@ -1,0 +1,17 @@
+import type { RefundPairRepository, TransferPairRepository } from '@budgie/contracts';
+
+export interface ConsolidationCandidateDependenciesInterface {
+    readonly refundPairRepository: Pick<RefundPairRepository, 'findCandidates' | 'findReviewCandidates'>;
+    readonly transferPairRepository: Pick<
+        TransferPairRepository,
+        | 'findAtmCashWithdrawalCandidates'
+        | 'findAtmCashWithdrawalReviewCandidates'
+        | 'findCandidates'
+        | 'findExistingTransferBridgeCandidates'
+        | 'findExistingTransferIncomeDuplicateCandidates'
+        | 'findIbanBridgeCanonicalDuplicateCandidates'
+        | 'findIbanBridgeChainTransferCandidates'
+        | 'findIbanBridgeTransferCandidates'
+        | 'findManualReviewCandidates'
+    >;
+}
