+++ packages/consolidation/src/auto/interface/consolidation-candidate-groups.interface.ts
