+++ packages/consolidation/src/auto/interface/consolidation-preview.interface.ts
