+++ packages/consolidation/src/auto/interface/consolidation-result.interface.ts
