+++ packages/consolidation/src/index.ts
@@ -1,1 +1,12 @@
+export { ConsolidationAutoCandidateService } from './auto/service/consolidation-auto-candidate.service';
+export { ConsolidationCandidateService } from './auto/service/consolidation-candidate.service';
+export { ConsolidationExecutorService } from './executor/service/consolidation-executor.service';
+
+export type { ConsolidationCandidateDependenciesInterface } from './auto/interface/consolidation-candidate-dependencies.interface';
+export type { ConsolidationCandidateGroupsInterface } from './auto/interface/consolidation-candidate-groups.interface';
+export type { ConsolidationPreviewInterface } from './auto/interface/consolidation-preview.interface';
+export type { ConsolidationProgressSnapshotInterface } from './auto/interface/consolidation-progress-snapshot.interface';
+export type { ConsolidationResultInterface } from './auto/interface/consolidation-result.interface';
+export type { CanonicalTransferInputInterface } from './executor/interface/canonical-transfer-input.interface';
+export type { ConsolidationExecutorDependenciesInterface } from './executor/interface/consolidation-executor-dependencies.interface';
 export type { TransactionRunnerInterface } from './wiring/interface/transaction-runner.interface';
